use atomic_add in cude scatter_elements. remove cuda implementation for scatter_add
